clean up errors/warnings from MSVC

+++ src/reader.c
@@ -27,6 +27,7 @@ The entry point is reader().  */
 
 #include <limits.h>
 #include <stdio.h>
+#include <string.h>
 #include <ctype.h>
 #include "system.h"
 #include "files.h"
@@ -469,7 +470,7 @@ parse_token_decl (what_is, what_is_not)
   register int token = 0;
   register int prev;
   register char *typename = 0;
-  int k;
+  size_t k;
 
 /*   start_lineno = lineno; JF */
 
@@ -545,7 +546,7 @@ parse_start_decl ()
 void
 parse_type_decl ()
 {
-  register int k;
+  register size_t k;
   register char *name;
 /*   register int start_lineno; JF */
 
@@ -602,7 +603,7 @@ void
 parse_assoc_decl (assoc)
 int assoc;
 {
-  register int k;
+  register size_t k;
   register char *name = NULL;
 /*  register int start_lineno; JF */
   register int prev = 0;	/* JF added = 0 to keep lint happy */
@@ -1596,7 +1597,7 @@ record_rule_line ()
 int
 get_type()
 {
-  register int k;
+  register size_t k;
   register int t;
   register char *name;
